firstofficer_service/customer-support-copilot/requirements.txt

numpy
python-dotenv
faiss-cpu
langchain-google-genai
numpy
sentence-transformers
langchain-community
pypdf